+++ wemake_python_styleguide/visitors/tokenize/primitives.py
@@ -23,6 +23,15 @@ def _replace_braces(string: str) -> str:
     return string.lstrip("'").rstrip("'")
 
 
+def _is_wrongly_underscored_number(string: str) -> bool:
+    parts = string.split('_')
+    if '.' in parts[-1]:
+        parts[-1] = parts[-1].split('.')[0]
+    is_correct_first_part = 1 <= len(parts[0]) <= 3
+    wrong_parts = [part for part in parts[1:] if len(part) != 3]
+    return not (is_correct_first_part and not wrong_parts)
+
+
 @final
 class WrongNumberTokenVisitor(BaseTokenVisitor):
     """Visits number tokens to find incorrect usages."""
@@ -55,7 +64,7 @@ class WrongNumberTokenVisitor(BaseTokenVisitor):
         self._check_float_zeros(token)
 
     def _check_underscored_number(self, token: tokenize.TokenInfo) -> None:
-        if '_' in token.string:
+        if '_' in token.string and _is_wrongly_underscored_number(token.string):
             self.add_violation(
                 consistency.UnderscoredNumberViolation(
                     token,